Major change to naming what is used
There are 4 Postgres flavors that can be used now + 1 for the
almost EOL BDR3.6, with the first argument:
	PG*
	PGE*
	3.6PGE*
	EPAS*
	BDRPG*
The second (optional) argument is the JIRA number (or any other
traking number), it creates a work directory with this name.
The third  (optional) argument is the extension version (BDR/PGL)

e.g.

`pgworkon PGE14 BDR-42 4`

This command will create a worktree in the $HOME/work directory
named `BDR-42` that will use Postgres Extended v14 and BDR v4.